Move getStaticCsvUrl and add to buildCache
